#1207 - groundspeed, alt and heading

+++ src/plugins/telemetry/include/plugins/telemetry/telemetry.h
@@ -650,10 +650,9 @@ public:
             float(NAN)}; /**< @brief Current indicated airspeed (IAS) in metres per second */
         float throttle_percentage{float(NAN)}; /**< @brief Current throttle setting (0 to 100) */
         float climb_rate_m_s{float(NAN)}; /**< @brief Current climb rate in metres per second */
-        float groundspeed_m_s{
-            float(NAN)}; /**< @brief Current indicated groundspeed in metres per second */
-        float alt_m{float(NAN)}; /**< @brief Current altitude (MSL)in metres */
-        uint32_t heading_deg{int16_t(NAN)}; /**< @brief Current heading in compass units (degrees) (0-360, 0=north).*/
+        float groundspeed_m_s{float(NAN)}; // Current indicated groundspeed in metres per second
+        float alt_m{float(NAN)}; // Current altitude (MSL) in metres
+        uint32_t heading_deg{uint32_t(NAN)}; // Current heading in compass units (0-360, 0=north)
     };
 
     /**
